requirements.txt
requests
tqdm
python-dotenv
GitPython
polars
numpy